fix bug

+++ TFT/src/User/API/interfaceCmd.c
@@ -186,11 +186,11 @@ void sendQueueCmd(void)
               }
             Serial_Puts(SERIAL_PORT_2, "Begin file list\n");
             if (mountFS() == true && scanPrintFiles() == true){
-              for (uint i = 0; i < infoFile.f_num; i++) {
+              for (uint16_t i = 0; i < infoFile.f_num; i++) {
                 Serial_Puts(SERIAL_PORT_2,infoFile.file[i]);
                 Serial_Puts(SERIAL_PORT_2,"\n");
                 } 
-              for (uint i = 0; i < infoFile.F_num; i++) {
+              for (uint16_t i = 0; i < infoFile.F_num; i++) {
                 Serial_Puts(SERIAL_PORT_2,"/");
                 Serial_Puts(SERIAL_PORT_2,infoFile.folder[i]);
                 Serial_Puts(SERIAL_PORT_2,"/\n");
